% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{CallCellRangerV2}
\alias{CallCellRangerV2}
\title{CellRanger_v2 method used to recognize cell-containing droplets.}
\usage{
CallCellRangerV2(
  counts,
  recovered_cells = 3000,
  recovered_cells_quantile = 0.99
)
}
\arguments{
\item{counts}{A \code{matrix} object or a \code{dgCMatrix} object which columns represent features and rows represent droplets.}

\item{recovered_cells}{Expected number of recovered cells. Default is 3000.}

\item{recovered_cells_quantile}{Quantile of the top \code{recovered_cells} barcodes by total UMI counts. Default is 0.99.}
}
\value{
A \code{DataFrame} containing the classification column named 'CellRangerV2Class'.
}
\description{
CellRanger_v2 method used to recognize cell-containing droplets.
}
\examples{
counts <- simSimpleCounts()
result <- CallCellRangerV2(counts)
head(result)
}
