% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{CallzUMIs}
\alias{CallzUMIs}
\title{zUMI method used to recognize cell-containing droplets.}
\usage{
CallzUMIs(counts)
}
\arguments{
\item{counts}{A \code{matrix} object or a \code{dgCMatrix} object which columns represent features and rows represent droplets.}
}
\value{
A \code{DataFrame} containing the classification column named 'zUMIsClass'.
}
\description{
zUMI method used to recognize cell-containing droplets.
}
\examples{
counts <- simSimpleCounts()
result <- CallzUMIs(counts)
head(result)
}
