% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{CellCalling}
\alias{CellCalling}
\title{Call cell-containg droplets using different methods.}
\usage{
CellCalling(counts, method = "dropSplit", seed = 0, ...)
}
\arguments{
\item{counts}{A \code{matrix} object or a \code{dgCMatrix} object which columns represent features and rows represent droplets.}

\item{method}{A cell-calling method from the following options:
\itemize{
\item \link[=dropSplit]{dropSplit}
\item \link[=CallCellRangerV2]{CellRangerV2}
\item \link[=CallCellRangerV3]{CellRangerV3}
\item \link[=CallEmptyDrops]{EmptyDrops}
\item \link[=CallzUMIs]{zUMIs}
}}

\item{seed}{Random seed used in the function. Default is 0.}

\item{...}{Arguments passed to the cell-calling method.}
}
\value{
A list including at least \code{DataFrame} named 'meta_info' which contains the final classification column named '\code{method}Class'.
}
\description{
Call cell-containg droplets using different methods.
}
\examples{
counts <- simSimpleCounts()
result <- CellCalling(counts, method = "dropSplit")
head(result)
}
