% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellEntropy.R
\name{CellEfficiency}
\alias{CellEfficiency}
\title{Calculate the entropy rate for the cells according to the feature counts.}
\usage{
CellEfficiency(x)
}
\arguments{
\item{x}{A sparse matrix which columns represent cells and rows represent features.}
}
\value{
A vector of the cell entropy rate.
}
\description{
Calculate the entropy rate for the cells according to the feature counts.
}
\examples{
x <- simCounts()
CER <- CellEfficiency(x)
head(CER)
}
