% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellEntropy.R
\name{CellEntropy}
\alias{CellEntropy}
\title{Calculate entropy for each cell according to the feature counts.}
\usage{
CellEntropy(x)
}
\arguments{
\item{x}{A sparse matrix which columns represent cells and rows represent features.}
}
\value{
A vector of the cell entropy.
}
\description{
Calculate entropy for each cell according to the feature counts.
}
\examples{
x <- simSimpleCounts()
CE <- CellEntropy(x)
head(CE)
}
