% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellEntropy.R
\name{CellEntropyRate}
\alias{CellEntropyRate}
\title{Calculate the entropy rate for the cells according to the feature counts.}
\usage{
CellEntropyRate(x)
}
\arguments{
\item{x}{A sparse matrix which columns represent cells and rows represent features.}
}
\value{
A vector of the cell entropy rate.
}
\description{
Calculate the entropy rate for the cells according to the feature counts.
}
\examples{
x <- DropletUtils:::simCounts()
CER <- CellEntropyRate(x)
head(CER)

}
