% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellGini.R
\name{CellGini}
\alias{CellGini}
\title{Calculate the Gini index for the cells according to the feature counts.}
\usage{
CellGini(x, normalize = TRUE)
}
\arguments{
\item{x}{A sparse matrix which columns represent cells and rows represent features.}

\item{normalize}{If \code{TRUE}, CellGini will add a number which represent the max Gini value when given the same counts.}
}
\value{
A vector of the cell entropy rate.
}
\description{
Calculate the Gini index for the cells according to the feature counts.
}
\examples{
x <- simSimpleCounts()
CG <- CellGini(x, normalize = TRUE)
head(CG)
}
