% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellEntropy.R
\name{CellRedundancy}
\alias{CellRedundancy}
\title{Calculate the Redundancy for the cells according to the feature counts.}
\usage{
CellRedundancy(x)
}
\arguments{
\item{x}{A sparse matrix which columns represent cells and rows represent features.}
}
\value{
A vector of the CellRedundancy.
}
\description{
Calculate the Redundancy for the cells according to the feature counts.
}
\examples{
x <- simCounts()
cr <- CellRedundancy(x)
head(cr)
}
