% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{compute_ambient_pvalues}
\alias{compute_ambient_pvalues}
\title{Compute p-values for observed multinomial log-likelihoods}
\usage{
compute_ambient_pvalues(umis_per_bc, obs_loglk, sim_n, sim_loglk)
}
\arguments{
\item{umis_per_bc}{UMI counts per barcode}

\item{obs_loglk}{Observed log-likelihoods of each barcode deriving from an ambient profile}

\item{sim_n}{Multinomial N for simulated log-likelihoods}

\item{sim_loglk}{Simulated log-likelihoods of maxtix. nrow=length(sim_n), ncol=num_simulations}
}
\value{
pvalues
}
\description{
Compute p-values for observed multinomial log-likelihoods
}
