% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{curvatureCalcluate}
\alias{curvatureCalcluate}
\title{Calculate curvature from the given x and y.}
\usage{
curvatureCalcluate(x, y)
}
\arguments{
\item{x, y}{A numeric vector.}
}
\value{
A list of the index of the raw x, y, and the corresponding first derivative, second derivative and curvature.
}
\description{
Calculate curvature from the given x and y.
}
\examples{
y <- c(0, 1, 3, 6, 9, 12, 11, 7, 9, 5, 1, 9, 0, 1, 2)
x <- 1:length(y)
result <- curvatureCalcluate(x = x, y = y)
result
}
