% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{est_background_profile_sgt}
\alias{est_background_profile_sgt}
\title{Estimate a gene expression profile on a given subset of barcodes.}
\usage{
est_background_profile_sgt(matrix, use_bcs)
}
\arguments{
\item{matrix}{Sparse matrix of all counts.}

\item{use_bcs}{Indices of barcodes to use.}
}
\value{
A list including:
\itemize{
\item \code{use_feats} use_features
\item \code{bg_profile_p} Estimated probabilities of length use_features.
}
}
\description{
Use Good-Turing to smooth the estimated profile.
}
