% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{estimate_profile_sgt}
\alias{estimate_profile_sgt}
\title{Estimate a gene expression profile by Simple Good Turing}
\usage{
estimate_profile_sgt(matrix, barcode_indices, nz_feat)
}
\arguments{
\item{matrix}{Sparse matrix of all counts}

\item{barcode_indices}{Barcode indices to use}

\item{nz_feat}{Indices of features that are non-zero at least once}
}
\value{
Estimated probabilities of length len(nz_feat).
}
\description{
Estimate a gene expression profile by Simple Good Turing
}
