% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{eval_multinomial_loglikelihoods}
\alias{eval_multinomial_loglikelihoods}
\title{Compute the multinomial log PMF for many barcodes}
\usage{
eval_multinomial_loglikelihoods(matrix, profile_p)
}
\arguments{
\item{matrix}{Matrix of UMI counts (feature x barcode)}

\item{profile_p}{Multinomial probability vector}
}
\value{
Log-likelihood for each barcode
}
\description{
Compute the multinomial log PMF for many barcodes
}
