% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_inflection}
\alias{find_inflection}
\title{Find the index of inflection in a numeric vector.}
\usage{
find_inflection(x, df = 20)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list of the index of the inflection and the corresponding value.
}
\description{
Find the index of inflection in a numeric vector.
}
\examples{
x <- DropletUtils:::simCounts()
inflection <- find_inflection(Matrix::colSums(x))
inflection
}
