% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{find_nonambient_barcodes}
\alias{find_nonambient_barcodes}
\title{Call barcodes as being sufficiently distinct from the ambient profile}
\usage{
find_nonambient_barcodes(
  matrix,
  orig_cell_bcs,
  n_partitions = 90000,
  n_candidate_barcodes = 20000,
  min_umis_nonambient = 500,
  min_umi_frac_of_median = 0.01,
  max_adj_pvalue = 0.01
)
}
\arguments{
\item{matrix}{Full expression matrix.}

\item{orig_cell_bcs}{Strings of initially-called cell barcodes.}

\item{n_partitions}{Number of partitions (max number of barcodes to consider for ambient estimation). Default is 90000.}

\item{n_candidate_barcodes}{Number of additional barcodes to consider after the initial cell calling. Default is 20000.}

\item{min_umis_nonambient}{Minimum number of UMIS per barcode to consider after the initial cell calling. Default is 500.}

\item{min_umi_frac_of_median}{Minimum ratio of UMIs to the median (initial cell call UMI) to consider after the initial cell calling. Default is 0.01.}

\item{max_adj_pvalue}{Minimum ratio of UMIs to the median (initial cell call UMI) to consider after the initial cell calling. Default is 0.01.}
}
\value{
A data.frame.
}
\description{
Call barcodes as being sufficiently distinct from the ambient profile
}
