% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peaks}
\alias{find_peaks}
\title{Find the index of peaks in a sequence of numeric values.}
\usage{
find_peaks(x, left_shoulder = 10000, right_shoulder = 10000)
}
\arguments{
\item{x}{A numeric vector.}

\item{left_shoulder}{A integer value.Peaks value should larger than any value in the left \code{left_shoulder} value.}

\item{right_shoulder}{A integer value.Peaks value should larger than any value in the right \code{right_shoulder} value.}
}
\value{
A vector of the index of peaks.
}
\description{
Find the index of peaks in a sequence of numeric values.
Peaks value should be larger than any values among the left/right shoulder.
If no peaks found, function will return the index of the max number.
}
\examples{
x <- c(0, 1, 3, 6, 9, 12, 11, 7, 9, 5, 1, 9, 0, 1, 2)
pks <- find_peaks(x, left_shoulder = 3, right_shoulder = 3)
pks
}
