% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellGini.R
\name{GiniScore}
\alias{GiniScore}
\title{Calculate GiniScore under a threshold.}
\usage{
GiniScore(x, GiniThreshold, group = NULL)
}
\arguments{
\item{x}{A vector of Gini index.}

\item{GiniThreshold}{A value used to estimate the feature-count redundancy from the Gini index. \code{x} larger than GiniThreshold will result in a GiniScore>0.5, else GiniScore<0.5.}

\item{group}{The group of \code{x}.}
}
\value{
A vector of the GiniScore.
}
\description{
Calculate GiniScore under a threshold.
}
\examples{
x <- c(0.6, 0.7, 0.8, 0.85, 0.9, 0.91, 0.92, 0.93, 0.95, 0.98, 0.99)
GiniScore(x, 0.95)
}
