% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ImportancePlot}
\alias{ImportancePlot}
\title{Plot the feature importance}
\usage{
ImportancePlot(meta_info, train, importance_matrix, top_n = 20)
}
\arguments{
\item{meta_info}{\code{importance_matrix} generated by dropSplit.}

\item{train}{\code{train} dataset generated by dropSplit.}

\item{importance_matrix}{\code{importance_matrix} generated by dropSplit.}

\item{top_n}{Number of the top important features to plot.}
}
\value{
A a list of ggplot objects.
}
\description{
Plot the feature importance
}
\examples{
counts <- simSimpleCounts()
result <- dropSplit(counts)
pl <- ImportancePlot(result$meta_info, result$train, result$importance_matrix, top_n = 20)
pl$Importance
pl$preDefinedClassExp
}
