% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outliers}
\alias{outliers}
\title{The IQR method of identifying outliers.}
\usage{
outliers(x, times = 1.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{times}{A constant used to discern bound.}
}
\value{
A list of the upper bound, the upper outliers index, the lower bound and the lower outliers index.
}
\description{
The IQR method of identifying outliers.
}
\examples{
x <- rep(seq(0.91, 1, 0.01), 2^(1:10))
boxplot(x)
outliers(x)
}
