% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{QCplot}
\alias{QCplot}
\title{QC plot for droplets using meta_info generated by dropSplit.}
\usage{
QCplot(
  meta_info,
  colorBy = "dropSplitClass",
  splitBy = "dropSplitClass",
  cell_stat_by = "dropSplitClass",
  metrics = c("nCount", "nFeature", "Rank", "RankMSE", "CellEntropy", "CellRedundancy",
    "CellGini"),
  ...
)
}
\arguments{
\item{meta_info}{\code{importance_matrix} generated by dropSplit.}

\item{colorBy}{One of the column names of the \code{meta_info} or NULL. Can be a type of continuous variable or discrete variable.}

\item{splitBy}{One of the column names of the \code{meta_info} or NULL. Must be a type of discrete variable.}

\item{cell_stat_by}{One of the discrete variable column names in \code{meta_info}. Used to count the number of 'Cell' in the column.}

\item{metrics}{QC metrics used to plot. Possible options are:
\itemize{
\item \code{nCount}
\item \code{nFeature}
\item \code{Rank}
\item \code{RankMSE}
\item \code{CellEntropy}
\item \code{CellRedundancy}
\item \code{CellGini}
}}

\item{...}{Other arguments passed to \code{\link{RankMSE}}.}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
QC plot for droplets using meta_info generated by dropSplit.
}
\examples{
counts <- simSimpleCounts()
result <- dropSplit(counts)
pl <- QCplot(result$meta_info, metrics = "CellEntropy")
pl$CellEntropy$Merge
pl$CellEntropy$Cell
}
