% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Score}
\alias{Score}
\title{Calculate Score under a threshold within groups.}
\usage{
Score(x, threshold, group = NULL, upper = NULL, lower = NULL)
}
\arguments{
\item{x}{A vector of metric used to score.}

\item{threshold}{A value used to calculate the score. \code{x} larger than threshold will result in a Score>0.5, else Score<0.5.}

\item{group}{The groups of \code{x}. If \code{NULL}, elements in \code{x} will be treated as the same group.}

\item{upper, lower}{A named vector of value to specify the upper/lower value for each group. Default is to find the upper for each group automatically.}
}
\value{
A vector of the Score. A Score>0.5 represent that corresponding x is larger than the \code{threshold} when fuzz=\code{FALSE}.
}
\description{
Calculate Score under a threshold within groups.
}
\examples{
x <- c(0.6, 0.7, 0.8, 0.85, 0.9, 0.91, 0.92, 0.93, 0.95, 0.98, 0.99)
Score(x, 0.95)
}
