% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{sgt_proportions}
\alias{sgt_proportions}
\title{Use Simple Good-Turing estimate to adjust for unobserved items}
\usage{
sgt_proportions(frequencies)
}
\arguments{
\item{frequencies}{Nonzero frequencies of items}
}
\value{
A list including:
\itemize{
\item \code{pstar} The adjusted non-zero proportions
\item \code{p0} The total probability of unobserved items
}
}
\description{
Use Simple Good-Turing estimate to adjust for unobserved items
}
