% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simCounts}
\alias{simCounts}
\title{Simulate counts for the empty and cell-containing droplets.}
\usage{
simCounts(
  ngenes = 5000,
  nempty = 20000,
  nlarge = 2000,
  nsmall = 200,
  empty.prof = seq_len(ngenes),
  empty.rate = 0.04,
  large.prof = empty.prof,
  large.rate = 0.01,
  large.shape = 10,
  small.prof = runif(ngenes),
  small.rate = 0.1,
  small.shape = 20,
  RemoveZeroCol = TRUE
)
}
\arguments{
\item{ngenes}{Total gene number for all the simulated counts.}

\item{nempty, nlarge, nsmall}{Empty, large cell and small cell droplets number.}

\item{empty.prof, large.prof, small.prof}{The overall gene expression profile distribution in distinct type of droplets.}

\item{empty.rate}{rate parameters in the \code{\link[stats]{Exponential}} function for empty droplets simulation.}

\item{large.rate, small.rate}{rate parameters in the \code{\link[stats]{GammaDist}} for large or small cell simulation.}

\item{large.shape, small.shape}{shape parameters in the \code{\link[stats]{GammaDist}} for large or small cell simulation.}

\item{RemoveZeroCol}{Whether to remove all zero-valued columns.}
}
\value{
A sparse Matrix of class "dgCMatrix".
}
\description{
Simulate counts for the empty and cell-containing droplets.
}
\examples{
counts <- simCounts()
counts
}
