% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{simple_good_turing}
\alias{simple_good_turing}
\title{Make a Simple Good-Turing estimate of the frequencies.}
\usage{
simple_good_turing(xr, xnr)
}
\arguments{
\item{xr}{Non-zero item frequencies}

\item{xnr}{Non-zero frequencies of frequencies}
}
\value{
A list including:
\itemize{
\item \code{rstar} The adjusted non-zero frequencies
\item \code{p0} The total probability of unobserved items
}
}
\description{
Make a Simple Good-Turing estimate of the frequencies.
}
