% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCalling.R
\name{simulate_multinomial_loglikelihoods}
\alias{simulate_multinomial_loglikelihoods}
\title{Simulate draws from a multinomial distribution for various values of N.}
\usage{
simulate_multinomial_loglikelihoods(
  profile_p,
  umis_per_bc,
  num_sims = 1000,
  jump = 1000,
  n_sample_feature_block = 1e+06,
  verbose = FALSE
)
}
\arguments{
\item{profile_p}{Probability of observing each feature.}

\item{umis_per_bc}{UMI counts per barcode (multinomial N).}

\item{num_sims}{Number of simulations per distinct N value.}

\item{jump}{Vectorize the sampling if the gap between two distinct Ns exceeds this.}

\item{n_sample_feature_block}{Vectorize this many feature samplings at a time.}

\item{verbose}{verbose. Default is FALSE.}
}
\value{
A list including:
\itemize{
\item \code{distinct_ns} an array containing the distinct N values that were simulated.
\item \code{log_likelihoods} a length(distinct_ns) x num_sims matrix containing the simulated log likelihoods.
}
}
\description{
Uses the approximation from Lun et al. ( https://www.biorxiv.org/content/biorxiv/early/2018/04/04/234872.full.pdf )
}
\examples{
profile_p <- c(.1, .2, .3, .1, .1, .2)
umis_per_bc <- c(100, 100, 200, 300)
simulate_multinomial_loglikelihoods(profile_p, umis_per_bc)
}
