% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{QCPlot}
\alias{QCPlot}
\title{QC plot for droplets using meta_info generated by dropSplit.}
\usage{
QCPlot(
  meta_info,
  colorBy = "dropSplitClass",
  splitBy = "dropSplitClass",
  metrics = c("nCount", "nFeature", "Rank", "RankMSE", "CellEntropy", "CellEfficiency",
    "CellGini")
)
}
\arguments{
\item{meta_info}{\code{importance_matrix} generated by dropSplit.}

\item{colorBy}{One of the column names of the \code{meta_info} or NULL. Can be a type of continuous variable or discrete variable.}

\item{splitBy}{One of the column names of the \code{meta_info} or NULL. Must be a type of discrete variable.}

\item{metrics}{QC metrics used to plot. Possible options are:
\itemize{
\item \code{nCount}
\item \code{nFeature}
\item \code{Rank}
\item \code{RankMSE}
\item \code{CellEntropy}
\item \code{CellEfficiency}
\item \code{CellGini}
}}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
QC plot for droplets using meta_info generated by dropSplit.
}
\examples{
counts <- simCounts()
result <- dropSplit(counts)
pl <- QCPlot(result$meta_info, metrics = "CellEntropy")
pl$CellEntropy$Merge
pl$CellEntropy$Cell
}
