% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropSplit.R
\name{RankMSE}
\alias{RankMSE}
\title{Calculate Mean Squared Error for nCount/nFeature Rank.}
\usage{
RankMSE(
  meta_info,
  fill_RankMSE = FALSE,
  smooth_num = 2,
  smooth_window = 100,
  find_cell_count = FALSE,
  tolerance = 0.2
)
}
\arguments{
\item{meta_info}{A \code{data.frame} or \code{DataFrame}. Must contain the 'nCount' and 'nFeature' columns.}

\item{fill_RankMSE}{Whether to fill the RankMSE by nCount. Default is \code{TRUE}.}

\item{smooth_num}{Number of times to smooth(take a mean value within a window length \code{smooth_window}) the squared error. Default is 2.}

\item{smooth_window}{Window length used to smooth the squared error. Default is 100.}

\item{find_cell_count}{Whether to find RankMSE valleys ('Cell' count cutoff). Default is FALSE.}

\item{tolerance}{A value indicated the tolerance when finding RankMSE valleys. A value greater than 1 indicates relaxed and will find more valleys; lower than 1 indicates strict and will find less valleys. Default is 0.2.}
}
\value{
A list include \code{meta_info} and \code{cell_rank_count}
}
\description{
Calculate Mean Squared Error for nCount/nFeature Rank.
}
