% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropSplit.R
\name{RankMSE}
\alias{RankMSE}
\title{Calculate Mean Squared Error for nCount/nFeature Rank.}
\usage{
RankMSE(
  meta_info,
  fill_RankMSE = FALSE,
  smooth_num = 2,
  smooth_window = 100,
  find_rank = FALSE,
  Empty_min_nCount = 10
)
}
\arguments{
\item{meta_info}{A \code{data.frame} or \code{DataFrame}. Must contain the 'nCount' and 'nFeature' columns.}

\item{fill_RankMSE}{Whether to fill the RankMSE by nCount. Default is \code{TRUE}.}

\item{smooth_num}{Number of times to smooth(take a mean value within a window length \code{smooth_window}) the squared error. Default is 2.}

\item{smooth_window}{Window length used to smooth the squared error. Default is 100.}

\item{find_rank}{Whether to find the 'Cell' RankMSE valley, the 'Uncertain' RankMSE peak and the 'Empty' RankMSE valley. Default is FALSE.}

\item{Empty_min_nCount}{Minimum nCount for 'Empty' droplets. Default is 10.}
}
\value{
A list include \code{meta_info} and \code{cell_rank_count}
}
\description{
Calculate Mean Squared Error for nCount/nFeature Rank.
}
